:recycle: refact: create files to put the coordinator protocols
