+++ controlsy/Sources/Core/Coordinators/MainCoordinator.swift
@@ -9,17 +9,6 @@ import Foundation
 import UIKit
 import RxSwift
 
-protocol CoordinatorProtocol {
-    var childCoordinators: [ChildCoordinatorProtocol] { get }
-    var navigation: UINavigationController { get }
-    
-    func start() -> UINavigationController
-    func getNavigationType() -> NavigationType
-    func addChildCoordinator(with coordinator: ChildCoordinatorProtocol)
-    func removeChildCoordinator(id: UUID)
-    func handleNavigation(to screen: NavigationType, navigation: UINavigationController)
-}
-
 enum NavigationType {
     case onboarding
     case session
